+++ inventory.py
@@ -3,18 +3,18 @@ import boto3
 from botocore.exceptions import ClientError
 import botocore
 import collections
-import datetime
-import time
-import dateutil
-from dateutil.tz import tzutc
 import csv
 import json
-from time import gmtime, strftime
+
 import smtplib
 import os, hmac, hashlib
 import pprint
 import logging
 from sys import exit
+import time
+from time import gmtime, strftime
+
+import res.internal_functions as internal
 
 # AWS Services imports 
 import res.ec2 as ec2
@@ -22,90 +22,66 @@ import res.s3 as s3
 import res.eks as eks
 import res.awslambda as awslambda
 
-#
-#  Useful functions
-#
-
-def datetime_converter(dt):
-    """
-        Converts a python datetime object (returned by AWS SDK) into a readable and SERIALIZABLE string
-
-        :param dt: datetime
-        :type region: datetime
-
-        :return: datetime in a good format
-        :rtype: str
-    """
-    if isinstance(dt, datetime.datetime):
-        return dt.__str__()  
-
-
-def json_datetime_converter(json_text):
-    """
-        Parses a json object and converts all datetime objects (returned by AWS SDK) into str objects
-
-        :param json_text: json with datetime objects
-        :type json_text: json
-
-        :return: json with date in string format
-        :rtype: json
-    """
-    return json.dumps(json_text, default = datetime_converter)      
-
-
-def get_ownerID():
-    """
-        Get owner ID of the AWS account we are working on
-
-        :return: owner ID
-        :rtype: string
-    """   
-    sts = boto3.client('sts')
-    identity = sts.get_caller_identity()
-    ownerId = identity['Account']
-    return ownerId
 
 #
 # Environment Variables & File handling & logging
 #
 
+# -- Some globals to initialize
+def init_globals():
+   
+    global filepath
+    global filename_json
+    global logger
+
+    # -- Some initial values
+    ownerId = internal.get_ownerID()
+
+    # --- Initial values for inventory files names
+    t = gmtime()
+    timestamp = strftime("%Y%m%d%H%M%S", t)
+    filepath = './output/'
+    filename_json = 'AWS_{}_{}.json'.format(ownerId, timestamp)
+
+    # --- logging variables
+    log_filepath    = './log/'
+    logger          = logging.getLogger('aws-inventory')
+    hdlr            = logging.FileHandler(log_filepath+'inventory.log')
+    formatter       = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
+
+    # --- Log handler
+    hdlr.setFormatter(formatter)
+    logger.addHandler(hdlr) 
+    logger.setLevel(logging.WARNING)
+
+    # --- If needed: S3 bucket name to write inventory
+    S3_INVENTORY_BUCKET="xx"
+
+#
+# Let's rock'n roll
+#
+init_globals()
+
 # --- AWS basic information
-ownerId = get_ownerID()
+ownerId = internal.get_ownerID()
 
 # --- AWS Regions 
 with open('aws_regions.json') as json_file:
     aws_regions = json.load(json_file)
 regions = aws_regions.get('Regions',[] ) 
 
-# --- Initial values for inventory files names
-t = gmtime()
-timestamp = strftime("%Y%m%d%H%M%S", t)
-filepath = './output/'
-filename_json = 'AWS_{}_{}.json'.format(ownerId, timestamp)
-# --- logging variables
-log_filepath = './log/'
-logger       = logging.getLogger('aws-inventory')
-hdlr         = logging.FileHandler(log_filepath+'inventory.log')
-formatter    = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
-# --- Log handler
-hdlr.setFormatter(formatter)
-logger.addHandler(hdlr) 
-logger.setLevel(logging.WARNING)
-# --- csv_file = open(filepath+filename,'w+')logging.basicConfig(filename='example.log',level=logging.DEBUG)
-S3_INVENTORY_BUCKET="xx"
 
-
-# Initialization for some variables
+# Initialization
 inventory = {}
-ec2_inventory        = []
-interfaces_inventory = []
-vpcs_inventory       = []
-ebs_inventory        = []
-lambda_inventory     = []
 
 # 
 # ----------------- EC2
 #
+
+ec2_inventory        = []
+interfaces_inventory = []
+vpcs_inventory       = []
+ebs_inventory        = []
 
 # Lookup in every AWS Region
 for current_region in regions:
@@ -113,18 +89,15 @@ for current_region in regions:
     current_region_name = current_region['RegionName']
     print('OwnerID : {}, EC2 inventory, Region : {}'.format(ownerId, current_region_name))
 
-    # Lambda
-    lambda_list = awslambda.get_lambda_inventory(ownerId, current_region_name)
-
-    # EC2
+    # EC2 instances
     instances = ec2.get_ec2_inventory(current_region_name)
     for instance in instances:
-       json_ec2_desc = json.loads(json_datetime_converter(instance))
+       json_ec2_desc = json.loads(internal.json_datetime_converter(instance))
        ec2_inventory.append(ec2.get_ec2_analysis(json_ec2_desc, current_region_name))
 
     # Network
     for ifc in ec2.get_interfaces_inventory(current_region_name):
-        interfaces_inventory.append(json.loads(json_datetime_converter(ifc)))
+        interfaces_inventory.append(json.loads(internal.json_datetime_converter(ifc)))
 
     # VPCs
     vpcs_inventory.append(ec2.get_vpc_inventory(current_region_name))
@@ -132,7 +105,7 @@ for current_region in regions:
     # EBS
     ebs_list = ec2.get_ebs_inventory(current_region_name)
     if len(ebs_list) > 0:
-        ebs_inventory.append(json.loads(json_datetime_converter(ebs_list)))
+        ebs_inventory.append(json.loads(internal.json_datetime_converter(ebs_list)))
 
     # EBS, snapshot
     # describe_nat_gateways()
@@ -145,6 +118,11 @@ inventory["ec2"]            = ec2_inventory
 inventory["ec2-interfaces"] = interfaces_inventory
 inventory["ec2-vpcs"]       = vpcs_inventory
 inventory["ec2-ebs"]        = ebs_inventory
+
+# 
+# ----------------- Lambda functions
+#
+inventory["lambda"]        = awslambda.get_lambda_inventory(ownerId)
 
 
 #
